split MINIT for Options
